move pykafka.producer __init__ argspec to base producer
